extract out withInnerHtmlFromUrl
